+++ kadai2/ebiiim/pkg/cli/cli_test.go
@@ -1,8 +1,9 @@
-package main
+package cli_test
 
 import (
 	"testing"
 
+	"github.com/gopherdojo/dojo5/kadai2/ebiiim/pkg/cli"
 	"github.com/gopherdojo/dojo5/kadai2/ebiiim/pkg/dirconv"
 	"github.com/gopherdojo/dojo5/kadai2/ebiiim/pkg/img"
 )
@@ -16,13 +17,13 @@ func check(t *testing.T, got interface{}, want interface{}) {
 func TestParseArgs(t *testing.T) {
 	// normal: no options
 	args := []string{"imgconv", "../../testdata"}
-	dc, _ := ParseArgs(args)
+	dc, _ := cli.ParseArgs(args)
 	check(t, *dc, dirconv.DirConv{Dir: "../../testdata", SrcExt: img.JPEG, TgtExt: img.PNG})
 }
 
 func TestParseArgs2(t *testing.T) {
 	// normal: with options
 	args := []string{"imgconv", "-source_ext=PNG", "-target_ext=.tiff", "../../testdata"}
-	dc, _ := ParseArgs(args)
+	dc, _ := cli.ParseArgs(args)
 	check(t, *dc, dirconv.DirConv{Dir: "../../testdata", SrcExt: img.PNG, TgtExt: img.TIFF})
 }
